passage de float a uint16_t pour la valeur de c02 (driver+bluetooth) + ajout d'un compteur permettant de savoir si le SCD41 n'a pas envoyé de données depuis 4 essaies (4 * 5 secondes = 20 secondes) + les flags (mouvement, son, obstacle) sont envoyés dans un seul octet dans le bluetooth

+++ src/core/measurement.hpp
@@ -6,7 +6,7 @@ struct Measurement
 {
     float temperature = NAN;
     float humidity    = NAN;
-    float co2         = NAN;
+    uint16_t co2      = NAN;
     float luminosity  = NAN;
     bool  motion      = false;
     bool  sound       = false;
